Split code_800D31A0.c into sys_freeze.c and sys_debug_controller.c
